break cli commands into separate modules
